refactor: move component oauth to component
